Use vtable for function pointers
