+++ src/lib.rs
@@ -1,7 +1,9 @@
+mod db;
 mod error;
 mod memory;
 
 pub use memory::{free_rust, Buffer};
+pub use crate::db::{db_t, DB};
 
 use crate::error::{handle_c_error, set_error};
 use std::panic::catch_unwind;
@@ -43,44 +45,6 @@ fn do_may_panic(guess: i32) -> Result<String, String> {
         Err("Too low".to_owned())
     } else {
         Ok("You are a winner!".to_owned())
-    }
-}
-
-// this represents something passed in from the caller side of FFI
-#[repr(C)]
-pub struct db_t { }
-
-#[repr(C)]
-pub struct DB {
-    pub state: *mut db_t,
-    pub c_get: extern fn(*mut db_t, Buffer, Buffer) -> i64,
-    pub c_set: extern fn(*mut db_t, Buffer, Buffer),
-}
-
-impl DB {
-    pub fn get(&self, key: Vec<u8>) -> Option<Vec<u8>> {
-        let buf = Buffer::from_vec(key);
-        // TODO: dynamic size
-        let mut buf2 = Buffer::from_vec(vec![0u8; 2000]);
-        let res = (self.c_get)(self.state, buf, buf2);
-
-        // read in the number of bytes returned
-        if res < 0 {
-            // TODO
-            panic!("val was not big enough for data");
-        }
-        if res == 0 {
-            return None
-        }
-        buf2.len = res as usize;
-        unsafe { Some(buf2.consume()) }
-    }
-
-    pub fn set(&self, key: Vec<u8>, value: Vec<u8>) {
-        let buf = Buffer::from_vec(key);
-        let buf2 = Buffer::from_vec(value);
-        // caller will free input
-        (self.c_set)(self.state, buf, buf2);
     }
 }
